+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -3,7 +3,7 @@ package com.techcourse;
 import jakarta.servlet.ServletContext;
 import nextstep.mvc.AnnotationHandlerAdapter;
 import nextstep.mvc.DispatcherServlet;
-import nextstep.mvc.ManualHandlerAdapter;
+import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
 import nextstep.web.WebApplicationInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -15,14 +15,27 @@ public class AppWebApplicationInitializer implements WebApplicationInitializer {
     @Override
     public void onStartup(final ServletContext servletContext) {
         final var dispatcherServlet = new DispatcherServlet();
-        dispatcherServlet.addHandlerMapping(new ManualHandlerMapping());
-        dispatcherServlet.addHandlerAdapter(new ManualHandlerAdapter());
-        dispatcherServlet.addHandlerAdapter(new AnnotationHandlerAdapter());
+        initStrategies(dispatcherServlet);
 
         final var dispatcher = servletContext.addServlet("dispatcher", dispatcherServlet);
         dispatcher.setLoadOnStartup(1);
         dispatcher.addMapping("/");
 
         log.info("Start AppWebApplication Initializer");
+    }
+
+    private void initStrategies(final DispatcherServlet dispatcherServlet) {
+        initHandlerMappings(dispatcherServlet);
+        initHandlerAdapters(dispatcherServlet);
+    }
+
+    private void initHandlerMappings(final DispatcherServlet dispatcherServlet) {
+        dispatcherServlet.addHandlerMapping(new ManualHandlerMapping());
+        dispatcherServlet.addHandlerMapping(new AnnotationHandlerMapping("package-name"));
+    }
+
+    private void initHandlerAdapters(final DispatcherServlet dispatcherServlet) {
+        dispatcherServlet.addHandlerAdapter(new ManualHandlerAdapter());
+        dispatcherServlet.addHandlerAdapter(new AnnotationHandlerAdapter());
     }
 }
