+++ app/src/test/java/com/techcourse/ManualHandlerAdapterTest.java
@@ -1,4 +1,4 @@
-package nextstep.mvc;
+package com.techcourse;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
@@ -10,7 +10,6 @@ import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import samples.TestController;
 
 class ManualHandlerAdapterTest {
